+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+### Added
+- Added an option to add `exportedWith` metadata to generated schemas
+  - This includes the game version and the Codec2Schema version
+
 ### Fixed
-- Fixed `cow_sound_variant` not being exported (oops)
+- Using `Utils#getPossibleValues` on an `EitherCodec` now works properly
+- Fixed default values in the config